arc: Workaround a possible issue with reading DDR memory
Similar to previous STAR, sometimes we read 0 instead of real memory
content, only this time it is a first read from memory, so we cannot add
artificial address writes. Instead now OpenOCD will check for JTAG status
register and will spin on this until it will set RD (ready) bit. In practice
it almost always set and I've never seen it not being set _before_ doing
memory read. So that is kind of a dummy useless check, since I haven't seen
cases where we would have to really wait for the JTAG interface. That
however seems to resolve the problem with failed reads, perhaps by adding a
delay between cache flush and memory read. Though that is quite strange
since in terms of the core time, there is already significant time between
flush and read.

Like with extra address writes this is applied only to "slow" memory (DDR)
and not to CCMs. It is also applied only to memory reads, though in theory
we should do this before each operations, that would add additional
performance penalty, and I don't feel inclined to decrease performance until
it will be obvious that it is required.

Signed-off-by: Anton Kolesov <Anton.Kolesov@synopsys.com>