tcl/target: Add reset handling for TI CC32xx
The CC3220/CC3235 SoC family has a restricted set of ways to perform a
reset, either at SoC or CPU level. The only method that leaves the
JTAG/SWD connection attached is the AIRCR VECTRESREQ bit. This change
to the TCL .cfg script for CC3220 adopts the 'cortex_m reset_config
vectreset' method to handle the reset command.  The reset is caught
(halted) by using the DEMCR debug feature to stop on fetch of reset
vector. After halting, the user reset vector and stack pointer are
read from the fixed address (as defined by SoC) and used to
initialize MSP and PC. This is similar to psoc4.cfg implementation.
Note that this only resets the CPU - not peripherals.

Signed-off-by: Mark Grosen <mgrosen@ti.com>
Signed-off-by: Kumar Gala <kumar.gala@linaro.org>